Refactor `select_conceptart` to `collect_conceptart` and ensure PEP08
